DeskCalc: Resize a bit more efficiently.
Remove B_FULL_UPDATE_ON_RESIZE flag from view.

We already Invalidate() on FrameResized() which redraws the
view on resize for us. This reduces the number of times Draw() is called
from 3+ in some cases to just one. Unfortunately Draw() is still called
multiple times in some cases producing noticable flickering and drawing
artifacts but it is marginally better.