Debugger: Implement new expression operands.
DWARF4 adds two new expression operands for values that are known, but
not located anywhere in memory (e.g. a known constant or a value computed
from a composite of other values/locations in the program). These are
DW_OP_implicit_value, which supplies the value directly as an LEB128 +
data block, and DW_OP_stack_value, which indicates the top of the
expression stack is the value.

Implements corresponding support in DwarfExpressionEvaluator,
ValueLocation and ValueLoader.