Debugger: Handle some more v4 attributes.
- DIECompileUnit can also be marked with DW_AT_main_subprogram to
indicate that the main program resides within that particular unit.

- DIEBaseType and DIEMember now have the additional attribute
DW_AT_data_bit_offset which defines the offset in a somewhat different
manner than DWARF2/3's DW_AT_bit_offset, and deprecates the latter.

Unused/untested for the moment, as gcc 4.7.3 doesn't yet appear to be
generating either of the above attributes.