Debugger: Fix handling of modified types with no base type.
In the case of a modified type pointing to an unspecified base type,
i.e. const void*, gcc appears to omit the DW_AT_type attribute entirely.
This had the result that such variables wouldn't appear in the var list.
If we encounter such a case, simply assume that the base type is unspecified
rather than failing early.