ScreenSaver: Refactor list item selection
Split out the list emptying into an EmptyScreenSaverList() method and use it in
various places where it should happen.

Call PopulateScreenSaverList() in ModulesView::AttachedToWindow() instead
of in the constructor because some of the methods depend on the view being
attached.

Don't try to select Blackness saver by name, instead if no other selection
is located select it as the default. Also, a tiny optimization, once we've found a
selection there's no reason to continue to keep looking for another one.