ScreenSaver: Open screensaver initially via message
..instead of calling _OpenSaver() directly.

At least one screensaver, Message, depended on the Bounds() of the
preview to be set when run, which wasn't established when you first
started the app causing the app to quit while trying to draw.

Also call PopulateSaverList() in AllAttached()  so that ScrollToSelection()
works without having to be called again.

Also, fix pointer style in ScreenSaverWindow::MessageReceived()