DeskBar: refactor expander code.
With the Tracker thread code in shared now it is possible to use the same
trick for the expanders in Deskbar to open and close in a separate thread
that we use in Tracker. See Find Panel and Info window for examples
of the code being used in Tracker.

Also eliminates the fClickedExpander variable and allow you to expand with any
mouse button.

Hopefully fixes #9676