debug_server: Add default action support.
- debug_server now supports user configuration of both the general default
action to be taken on crashes, and per-app overrides. These are stored
config/settings/system/debug_server/settings. In the absence of the file,
behavior defaults to prompting the user for an action as before.

The file takes on a driver_settings style format as follows:

---
default_action user

executable_actions {
	app1 log
	/path/app2* debug
}
---

Valid values for an action are:
	- user = prompt user for action.
	- kill = silently terminate the crashing team.
	- debug = attach the debugger to the crashing team. in the debugger.
	- log / report = save a crash report and terminate the team.

If no default_action is specified, "user" is assumed.

The executable_actions subsection contains individual overrides of the
default. These take the form above, where the individual lines can be
only a team name, or a path, with wildcards.

Implements first part of #10301.